Make multiple components importable #26
